The presence of a number in a cell
The algorithm passes through the traced array element by element. The order of passing through the array depends on the direction of movement (or the pressed button).
If there is a number in the cell, then the internal logic will be executed. Otherwise, the number of empty cells will be incremented.