Number of empty cells
We go through the row or column element by element, depending on the direction of the shift. Empty cells are counted in the opposite direction of the shift. The number of empty cells is necessary to calculate how much it is necessary to shift this or that number in a given direction.